fix: improve token validation
