+++ src/navigator/Navigator.tsx
@@ -10,12 +10,7 @@ import { CreateIdeaScreen } from '../screens/CreateIdeaScreen';
 import { OpenedIdeaScreen } from '../screens/OpenedIdeaScreen';
 import { ManifestPart1Screen } from '../screens/ManifestPart1Screen';
 import { RootState } from '../store';
-import { useAppDispatch, useAppSelector } from '../store/hooks';
-import SpikyService from '../services/SpikyService';
-import { setUniversities } from '../store/feature/ui/uiSlice';
-import { University, User } from '../types/store';
-import { addToast } from '../store/feature/toast/toastSlice';
-import { StatusType } from '../types/common';
+import { useAppSelector } from '../store/hooks';
 import { TermAndConditionsScreen } from '../screens/TermAndConditionsScreen';
 import { ReportIdeaScreen } from '../screens/ReportIdeaScreen';
 import { ReplyIdeaScreen } from '../screens/ReplyIdeaScreen';
@@ -23,6 +18,7 @@ import { ChatScreen } from '../screens/ChatScreen';
 import { ChangeForgotPasswordScreen } from '../screens/ChangeForgotPasswordScreen';
 import { ChangePasswordScreen } from '../screens/ChangePasswordScreen';
 import SocketContext from '../context/Socket/Context';
+import { User } from '../types/store';
 
 export type RootStackParamList = {
     HomeScreen: undefined;
@@ -53,34 +49,9 @@ export type RootStackParamList = {
 const Stack = createStackNavigator<RootStackParamList>();
 
 export const Navigator = () => {
-    const dispatch = useAppDispatch();
-    const token = useAppSelector((state: RootState) => state.auth.token);
     const config = useAppSelector((state: RootState) => state.serviceConfig.config);
-    const universities = useAppSelector((state: RootState) => state.ui.universities);
     const appState = useAppSelector((state: RootState) => state.ui.appState);
     const { socket } = useContext(SocketContext);
-
-    async function setSessionInfo() {
-        const spikyClient = new SpikyService(config);
-        try {
-            const { data: universitiesData } = await spikyClient.getUniversities();
-            const { universidades } = universitiesData;
-            const universitiesResponse: University[] = universidades.map<University>(
-                university => ({
-                    id: university.id_universidad,
-                    shortname: university.alias,
-                    color: university.color,
-                    backgroundColor: university.background_color,
-                })
-            );
-            dispatch(setUniversities(universitiesResponse));
-        } catch (e) {
-            console.log(e);
-            dispatch(
-                addToast({ message: 'Error cargando universidades', type: StatusType.WARNING })
-            );
-        }
-    }
 
     // I changed this because the token in store.auth can be
     // defined before config.headers.x-token that is the one
@@ -88,14 +59,9 @@ export const Navigator = () => {
     // TODO: centralize in one place where to put the token,
     // because i think it is saved in axios, in SecureStorage
     // and in store's auth.token
-    useEffect(() => {
-        if (config?.headers?.['x-token'] && !universities) {
-            setSessionInfo();
-        }
-    }, [universities, config]);
 
     useEffect(() => {
-        if (appState === 'inactive') {
+        if (appState === 'inactive' && config?.headers?.token) {
             socket?.emit('force-offline', {});
         } else {
             socket?.emit('force-online', {});
@@ -111,7 +77,7 @@ export const Navigator = () => {
                 },
             }}
         >
-            {!token ? (
+            {!config?.headers?.token ? (
                 <>
                     <Stack.Screen name="HomeScreen" component={HomeScreen} />
                     <Stack.Screen name="LoginScreen" component={LoginScreen} />
